[Refactoring] 포인트 double -> int 타입으로 변경 및 상품등록시 포인트 목록 int 타입으로 강제캐스팅

+++ src/main/java/com/dbzz/matchingproject/dto/response/AllProductResponseDto.java
@@ -11,7 +11,7 @@ public class AllProductResponseDto {
     private String productName;
     private int price;
     private String productInfo;
-    private double point;
+    private int point;
 
     public AllProductResponseDto(Product product) {
         this.userId = product.getUserId();
